requirements.txt
setuptools>=52.0.0
Django>=3.1.6
pytz>=2021.1
sqlparse>=0.4.1
asgiref>=3.3.1
djangorestframework>=3.12.2
pandas
django-taggit
geocoder
folium
Pillow
 oauthlib
 requests-oauthlib
 python3-openid
 PyJWT
 social-auth-core
 social-auth-app-django
 django-widget-tweaks